Topological sort before rewriting aggregate jobs
Given the leaf jobs `packages.foo` & `packages.bar`, an aggregate job
`aggregate0` with

    _hydraGlobConstituents = true;
    constituents = [ "packages.*" ];

and an aggregate job `aggregate1` with

    constituents = [ "aggregate0" ];

then it may happen depending on the order of evaluation that `aggregate1`
depends on the old derivation of `aggregate0` (i.e. the one without
rewritten constituents) and doesn't depend on `packages.foo` and
`packages.bar` because it was rewritten before `aggregate0` was
rewritten.

Using a toposort (derived from the one in CppNix's libutil) to solve
that. Cycles are a hard error, then _all_ aggregates are failed to make
sure we don't finish one of those too early.

This also extracts the code into a few helper functions and a slightly
better data structure, otherwise this would've gotten too chaotic.